its on the way (soon work) # 3 the backpack wont open

+++ src/main/java/ch/skyfy/enderbackpack/client/screen/BackpackScreen.java
@@ -14,7 +14,7 @@ import net.minecraft.util.Identifier;
 @Environment(EnvType.CLIENT)
 public class BackpackScreen extends HandledScreen<BackpackScreenHandler> {
 
-    private final Byte row;
+    private final int row;
 
     public BackpackScreen(BackpackScreenHandler handler, PlayerInventory inventory, Text title) {
         super(handler, inventory, title);
